Imperial Entanglements
Set.xml only for Imperial Entanglements - no scripts for any of the
cards yet.

I suggest this shouldn't be added to the game feed just yet since the
cards aren't up on CardgameDB, but this could be a head start to getting
the set live with at least basic scripts to follow.